speed up kiill_with_shell + minor changes

+++ home/ss/.config/awesome/rc.lua
@@ -691,8 +691,10 @@ local function spawn_with_shell_once(pname, cmd)
 end
 
 local function kill_with_shell(pname)
-    awful.spawn.with_shell("pkill -9 " .. pname)
-    os.execute("sleep 1")
+    if isrunning(pname) then
+        awful.spawn.with_shell("pkill -9 " .. pname)
+        os.execute("sleep 1")
+    end
 end
 
 local function respawn_with_shell(pname, cmd)
@@ -752,12 +754,12 @@ respawn_with_shell("xautolock", "xautolock -detectsleep -time 10 -notify 30 -not
 respawn_with_shell("wicd-client", "wicd-client --tray &")
 spawn_once("xpad")
 
+spawn_once("thunderbird")
+-- respawn("birdtray") -- TODO: doesn't work
 --spawn_once("davmail")
---spawn_once("thunderbird")
---respawn("birdtray") DOESN"T work good
-spawn_once("evolution")
-spawn_once("skypeforlinux")
-spawn_once("hipchat4")
+--spawn_once("evolution")
+--spawn_once("skypeforlinux")
+--spawn_once("hipchat4")
 
 spawn_once("pavucontrol")
 spawn_once("audacious")
